+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* Allow using the `sql_log_normalizer` Sequel database extension (@janko)
+
 ## 1.2.4 (2021-09-27)
 
 * Allow using with Active Record 7.0 (@janko)